unicorn-docs/requirements.txt
chromadb>=0.4.0
requests>=2.25.0
python-dotenv>=0.19.0
